+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes
 
 v1.8.1 (??.06.2022)
 * removed all anient model classes
+* Fix filter classes
+* Fix fatal error in stadium list
 
 v1.8.0 (19.06.2022)
 * all classes beside from models are moved to PSR-4 namespace